feat(demos): full curriculum alignment across all 9 modules
Each module's hands-on demo now exercises the topics promised by its
notion-guides/HTML curriculum, not just a tactical subset. Every module
gets an extras.sql (or in M8's case, extra drills inside run.sh) plus
README updates documenting the additions.

  M1  TTL · codecs (Delta/T64/ZSTD) · Array/Tuple/Map/Nested/Enum/Nullable
      DESCRIBE TABLE · HTTP API smoke test
  M2  VersionedCollapsingMergeTree · Materialized View pipeline
      Nested type with ARRAY JOIN
      (also fixes a pre-existing CH 26 strictness bug in the AggregatingMergeTree
      state and an argMax+max alias collision)
  M3  weighted shards (second cluster definition with weights 1,1,4)
      alternative sharding keys: rand() · intDiv() · xxHash64()
  M4  insert_quorum · select_sequential_consistency
      SYSTEM DROP REPLICA inspection · ClickHouse Keeper note
  M5  users.xml: 3 users / 3 profiles / 2 quotas
      SQL roles + grants on top
      Prometheus endpoint enabled (port 9363) + smoke-test
      TLS sketch
  M6  PREWHERE · SAMPLE BY · skip indexes (minmax/set/tokenbf_v1)
      JOIN strategies (ANY/ALL/Dictionary)
      Materialized View with backfill
  M7  async backup · incremental backup (base_backup) + RESTORE chain
      clickhouse-backup CLI ref · BACKUP ON CLUSTER note
  M8  drill 5: insert_quorum durability gate (1/2 replicas → fail)
      drill 6: BACKUP-RESTORE recovery path
      multi-DC architecture sketch in README
  M9  DLQ via kafka_handle_error_mode='stream' (parallel safe consumer)
      exactly-once via ReplacingMergeTree on (user_id, event_time)
      run.sh now produces 50 broken JSON lines after the valid batch
      consumer-group offset reset commands documented

Top-level README's module table is updated to summarise both the core
demo and the extras for each module.

All 9 compose files validate, every config XML is well-formed (xmllint),
shell scripts pass `bash -n`. M1 and M2 smoke-tested end-to-end on
clickhouse/clickhouse-server:latest (26.4.2.10).

+++ code-examples/demos/module-2-table-engines/setup.sql
@@ -37,7 +37,7 @@ CREATE TABLE m2.events_agg
     country         LowCardinality(String),
     uniq_users_state AggregateFunction(uniq, UInt32),
     revenue_state    AggregateFunction(sum, Float64),
-    p99_state        AggregateFunction(quantileTDigest(0.99), Float32)
+    p99_state        AggregateFunction(quantileTDigest(0.99), Float64)
 )
 ENGINE = AggregatingMergeTree
 ORDER BY (bucket_date, country);
